separate tree related typings
