+++ desci-models/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./ResearchObject";
 export * from "./constants";
 export * from "./transformers/index";
-export * from "./treeTools";
+export * from "./trees/treeTools";
+export * from "./trees/treeTypes";
